requirements.txt
gspread
google-auth
PyQt5
pyinstaller
